davici: Don't limit initially allocated buffer to 32 bytes
As noted in 75b1681064d9 ("davici: Skip superfluous max() check for
buffer increase") it is assumed that messages are at least 32 bytes
long.  This is not the case with min().  Instead it even limits the
length of the allocated buffer to 32 bytes if the passed name is
longer (not the case for the currently defined operations) causing
a buffer overflow in the memcpy() that follows, which is based on
the original length.

Reported-by: Stefan Berghofer <stefan.berghofer@secunet.com>