davici: Set O_NONBLOCK for non-blocking I/O instead of using MSG_DONTWAIT
While MSG_DONTWAIT is defined it does not seem to have any effect in
send/recv calls (it's actually not documented in the corresponding man
pages) on Mac OS X or FreeBSD.

We explicitly do this after connect(), as the constructor does not handle
asynchronous connects. This is mentioned in the documentation, and usually
makes not much sense for UNIX sockets.

Additionally, at the same time we add O_CLOEXEC, which is usually good
behavior.