davici: Cast isprint() argument to unsigned char
isprint() is invoked here with an argument of signed type char,
but only has defined behavior for int arguments that are either
representable as unsigned char or equal to the value of macro
EOF(-1).

As per codesonar report, in a number of libc implementations,
isprint() is implemented using lookup tables (arrays): passing
in a negative value can result in a read underrun.

To avoid this unexpected behaviour, typecasted char type
argument to unsigned char type.